Fix osscilation in Scroll (#1)
Rationale: GetSize is not a fixed value - it depends on actual state of frame. As we are using AutoHide scrollbar here, GetSize can depend on whether ScrollBar is visible or not, which can lead to infinite recursive loop.

The correct approach is to specify the size at AddPane and use GetLayoutSize there (maybe fix example to do this too); with fallback overload that just snapshots GetSize at the moment of AddPane - that is less safe, but should work in most situations.